Очищать прерванные worker-тесты вместе с fake gh
Контролируемый test-helper теперь удаляет временный корень и завершает собственную process group после TERM, INT и HUP. Добавлена регрессия для PID, группы и каталога; production-код не менялся. Проверено целевыми worker-тестами и полным go test ./.... Карточка CARD-0097.

+++ internal/worker/worker_integration_test.go
@@ -281,6 +281,10 @@ var repositoryTemplate struct {
 }
 
 func TestMain(main *testing.M) {
+	if os.Getenv(interruptedTestHelperEnv) == "1" {
+		runInterruptedTestHelper()
+		return
+	}
 	code := main.Run()
 	if repositoryTemplate.root != "" {
 		_ = os.RemoveAll(repositoryTemplate.root)
